[wasm] DebuggerTests - Split EvaluateOnCallFrameTests
